createResident updateResident getActiveResident moveOut

+++ src/main/java/com/ptithcm/apt/repository/ResidentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ptithcm.apt.entity.Resident;
 
-public interface ResidentRepository extends JpaRepository<Resident, Integer> {
+public interface ResidentRepository extends JpaRepository<Resident, Long> {
     boolean existsByCitizenIdentity(String citizenIdentity);
 
     Optional<Resident> findByCitizenIdentity(String citizenIdentity);
